fix: enforce skill discovery depth for dotdot names
